[BPK-4242] Add trailingAccessoryView prop to BpkChip
