+++ UNRELEASED.md
@@ -2,6 +2,10 @@
 
 > Place your changes below this line.
 
+**Added:**
+
+- Added `trailingAccessoryView` prop to `BpkChip`.
+
 
 ## How to write a good changelog entry
