+++ lib/bpk-component-chip/src/BpkChip-test.common.js
@@ -130,7 +130,7 @@ const commonTests = () => {
       expect(tree).toMatchSnapshot();
     });
 
-    it('should render correctly with a supplied icon', () => {
+    it('should render correctly with leadingAccessoryView', () => {
       const tree = renderer
         .create(
           <BpkChip
@@ -138,6 +138,21 @@ const commonTests = () => {
             label="Label"
             accessibilityLabel="Accessibility label"
             leadingAccessoryView={<BpkIcon icon={icons.flight} />}
+          />,
+        )
+        .toJSON();
+
+      expect(tree).toMatchSnapshot();
+    });
+
+    it('should render correctly with trailingAccessoryView', () => {
+      const tree = renderer
+        .create(
+          <BpkChip
+            onPress={onPress}
+            label="Label"
+            accessibilityLabel="Accessibility label"
+            trailingAccessoryView={<BpkIcon icon={icons.flight} />}
           />,
         )
         .toJSON();
