+++ lib/bpk-component-chip/src/BpkDismissibleChip.js
@@ -20,7 +20,9 @@
 
 import React from 'react';
 
-import BpkChipWrapper from './BpkChipWrapper';
+import BpkIcon, { icons } from '../../bpk-component-icon';
+
+import BpkChip from './BpkChip';
 import {
   type Props as CommonProps,
   commonPropTypes,
@@ -32,7 +34,14 @@ export type Props = CommonProps;
 const BpkDismissibleChip = (props: Props) => {
   const { ...rest } = props;
 
-  return <BpkChipWrapper {...rest} dismissible large selected={false} />;
+  return (
+    <BpkChip
+      {...rest}
+      trailingAccessoryView={<BpkIcon icon={icons.close} small />}
+      large
+      selected={false}
+    />
+  );
 };
 
 BpkDismissibleChip.propTypes = commonPropTypes;
