+++ lib/bpk-component-chip/src/common-types.js
@@ -19,7 +19,7 @@
 /* @flow */
 
 import PropTypes from 'prop-types';
-import { type Node, type ElementProps } from 'react';
+import { type Node, type ElementProps, type Element } from 'react';
 import { View, ViewPropTypes } from 'react-native';
 
 import { type Theme } from '../../bpk-theming';
@@ -41,6 +41,7 @@ export type Props = {
   label: string,
   style: ViewStyleProp,
   innerChipStyle: ViewStyleProp,
+  leadingAccessoryView: ?Element<*>,
   type: ?$Keys<typeof CHIP_TYPES>,
   theme: ?Theme,
 };
@@ -59,16 +60,18 @@ export const commonPropTypes = {
   disabled: PropTypes.bool,
   onPress: PropTypes.func.isRequired,
   label: PropTypes.string.isRequired,
-  style: ViewPropTypes.style,
   innerChipStyle: ViewPropTypes.style,
-  type: PropTypes.oneOf(Object.keys(CHIP_TYPES)),
+  leadingAccessoryView: PropTypes.element,
+  style: ViewPropTypes.style,
   theme: themePropType,
+  type: PropTypes.oneOf(Object.keys(CHIP_TYPES)),
 };
 
 export const commonDefaultProps = {
   disabled: false,
-  style: null,
   innerChipStyle: null,
-  type: CHIP_TYPES.primary,
+  leadingAccessoryView: null,
+  style: null,
   theme: null,
+  type: CHIP_TYPES.primary,
 };
