+++ lib/bpk-component-chip/stories.js
@@ -254,6 +254,24 @@ storiesOf('bpk-component-chip', module)
       </View>
     </View>
   ))
+  .add('docs:with-trailing-icon', () => (
+    <View style={styles.bottomMargin}>
+      <View style={styles.row}>
+        {COUNTRIES.map((country, index) => (
+          <BpkChip
+            key={country}
+            label={country}
+            accessibilityLabel={`Toggle ${country}`}
+            onPress={() => {}}
+            selected={index % 4 === 0}
+            disabled={index % 10 === 0}
+            style={styles.chip}
+            trailingAccessoryView={<BpkIcon icon={icons.flight} />}
+          />
+        ))}
+      </View>
+    </View>
+  ))
   .add('docs:outline', () => (
     <View style={[styles.bottomMargin, outlineStyles.row]}>
       <View style={styles.row}>
